common/saclient: import path fixup after submodule rename
