+++ common/saclient/endpoint_test.go
@@ -20,7 +20,7 @@ import (
 	"os"
 	"testing"
 
-	. "github.com/sacloud/sacloud-sdk-go/internal/saclient"
+	. "github.com/sacloud/sacloud-sdk-go/common/saclient"
 	"github.com/stretchr/testify/suite"
 )
