+++ common/saclient/error_test.go
@@ -17,7 +17,7 @@ package saclient_test
 import (
 	"testing"
 
-	. "github.com/sacloud/sacloud-sdk-go/internal/saclient"
+	. "github.com/sacloud/sacloud-sdk-go/common/saclient"
 	"github.com/stretchr/testify/suite"
 )
