+++ common/saclient/http_request_doer_test.go
@@ -25,8 +25,8 @@ import (
 	"strings"
 	"testing"
 
+	. "github.com/sacloud/sacloud-sdk-go/common/saclient"
 	"github.com/sacloud/sacloud-sdk-go/internal/packages/testutil"
-	. "github.com/sacloud/sacloud-sdk-go/internal/saclient"
 	"github.com/stretchr/testify/suite"
 )
