+++ common/saclient/profile_test.go
@@ -22,7 +22,7 @@ import (
 	"runtime"
 	"testing"
 
-	. "github.com/sacloud/sacloud-sdk-go/internal/saclient"
+	. "github.com/sacloud/sacloud-sdk-go/common/saclient"
 	"github.com/stretchr/testify/suite"
 )
