+++ tool/replace.sh
@@ -20,7 +20,7 @@ declare -A map=(
     ["nosql-api-go"]="api/nosql"
     ["object-storage-api-go"]="api/object-storage"
     ["packages-go"]="internal/packages"
-    ["saclient-go"]="internal/saclient"
+    ["saclient-go"]="common/saclient"
     ["secretmanager-api-go"]="api/secretmanager"
     ["security-control-api-go"]="api/security-control"
     ["service-endpoint-gateway-api-go"]="api/service-endpoint-gateway"
